% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearSchema.R
\name{clearSchema}
\alias{clearSchema}
\title{Clear the Current Target Schema}
\usage{
clearSchema(conn, targetSchema, drop = FALSE)
}
\arguments{
\item{conn}{A Connection object.}

\item{drop}{If TRUE, will drop the schema if it exists.}
}
\description{
Clear the Current Target Schema
}
