% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints}
\alias{constraints}
\title{Execute Athena Constraints}
\usage{
constraints(conn, targetSchema)
}
\description{
The constraints are performed using the SQL found at \href{https://raw.githubusercontent.com/OHDSI/CommonDataModel/master/PostgreSQL/OMOP\%20CDM\%20postgresql\%20constraints.txt}{OHDSI OMOP CDM Postgresl Constraints}
}
\seealso{
\code{\link[SqlRender]{render}},\code{\link[SqlRender]{readSql}}
\code{\link[pg13]{sourceFilePath}},\code{\link[pg13]{parseSQL}},\code{\link[pg13]{sendList}}
}
