% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMRCONSO.R
\name{createMRCONSO}
\alias{createMRCONSO}
\title{Add the MRCONSO UMLS Metathesaurus}
\usage{
createMRCONSO(conn, schema = "umls", drop_schema = TRUE)
}
\arguments{
\item{schema}{name of schema to write to}

\item{drop_schema}{If TRUE, will drop schema cascade before creating a new schema. If FALSE, the schema parameter must exist.}

\item{filePath}{path to RRF files}
}
\description{
Add the MRCONSO UMLS Metathesaurus
}
\seealso{
\code{\link[SqlRender]{render}},\code{\link[SqlRender]{readSql}}
\code{\link[pg13]{sourceFilePath}},\code{\link[pg13]{send}}
}
