% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUMLS.R, R/ddl.R
\name{addUMLS}
\alias{addUMLS}
\alias{ddl}
\title{Add the MRCONSO UMLS Metathesaurus}
\usage{
addUMLS(conn)

ddl(conn, targetSchema)
}
\description{
Add the MRCONSO UMLS Metathesaurus

Execute Athena DDL
}
\seealso{
\code{\link[SqlRender]{render}},\code{\link[SqlRender]{readSql}}
\code{\link[pg13]{sourceFilePath}},\code{\link[pg13]{send}}

\code{\link[SqlRender]{render}},\code{\link[SqlRender]{readSql}}
\code{\link[pg13]{sourceFilePath}},\code{\link[pg13]{send}}
}
