% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_last_log}
\alias{fetch_last_log}
\title{FUNCTION_TITLE}
\usage{
fetch_last_log(conn, conn_fun, verbose = FALSE, render_sql = FALSE)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
