% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_omop_release_version}
\alias{fetch_omop_release_version}
\title{FUNCTION_TITLE}
\usage{
fetch_omop_release_version(conn, conn_fun)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
