% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{get_github_release}
\alias{get_github_release}
\title{Get the most current OMOP Vocabulary release available}
\usage{
get_github_release(
  local_repo_path = "~/GitHub/OHDSI/Vocabulary-v5.0/",
  cdm_version = "V5.0"
)
}
\arguments{
\item{local_repo_path}{Path to the cloned OHDSI Vocabulary repository, Default: '~/GitHub/OHDSI/Vocabulary-v5.0/'}

\item{cdm_version}{String that is appended to the beginning of the tag used to determine the version, Default: 'V5.0'}
}
\description{
Get the most current OMOP Vocabulary release available
}
