% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{get_version}
\alias{get_version}
\title{Get the current OMOP Vocabulary version in the Database}
\usage{
get_version(
  conn,
  conn_fun = "pg13::local_connect()",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\description{
Get the current OMOP Vocabulary version in the Database
}
