% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{Execute Athena Indexes}
\usage{
indices(conn, target_schema, verbose = TRUE, render_sql = TRUE)
}
\description{
SQL for indices derived from \url{"https://raw.githubusercontent.com/OHDSI/CommonDataModel/master/PostgreSQL/OMOP\%20CDM\%20postgresql\%20pk\%20indexes.txt"}
}
\seealso{
\code{\link[SqlRender]{render}}
\code{\link[pg13]{execute_n}}
}
