% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMRCONSO.R
\name{loadMRCONSO}
\alias{loadMRCONSO}
\title{Load MRCONSO Table}
\usage{
loadMRCONSO(filePath, max_lines, conn, schema)
}
\arguments{
\item{filePath}{PARAM_DESCRIPTION}

\item{max_lines}{PARAM_DESCRIPTION}

\item{conn}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[progress]{progress_bar}}
\code{\link[readr]{read_delim}}
\code{\link[pg13]{appendTable}}
}
