% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log}
\alias{log}
\title{Log Update}
\usage{
log(conn, target_schema, verbose = TRUE, render_sql = TRUE, release_version)
}
\description{
This function prints the number of rows for all the
vocabulary tables in the R console.
}
