% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_args.R
\name{pkg_args}
\alias{pkg_args}
\title{setupAthena Arguments}
\arguments{
\item{umls_api_key}{PARAM_DESCRIPTION}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{conn}{PARAM_DESCRIPTION, Default: conn}

\item{target_schema}{PARAM_DESCRIPTION, Default: 'public'}

\item{cascade}{PARAM_DESCRIPTION, Default: TRUE}

\item{path_to_csvs}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{steps}{Options include "prepare_cpt4".}
}
\description{
setupAthena Arguments
}
