% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_setup.R
\name{run_setup}
\alias{run_setup}
\title{Run Athena Setup}
\usage{
run_setup(
  conn,
  conn_fun,
  target_schema = "omop_vocabulary",
  steps = c("drop_tables", "copy", "indices", "constraints", "log"),
  path_to_csvs,
  umls_api_key,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION, Default: conn}

\item{target_schema}{PARAM_DESCRIPTION, Default: 'public'}

\item{steps}{Options include "prepare_cpt4".}

\item{path_to_csvs}{PARAM_DESCRIPTION}

\item{umls_api_key}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
Updated OMOP Vocabulary (Athena) CONCEPT_ANCESTOR, CONCEPT_CLASS, CONCEPT_RELATIONSHIP, CONCEPT_SYNONYM, CONCEPT, DOMAIN, DRUG_STRENGTH, RELATIONSHIP, and VOCABULARY Tables in the given target schema.
}
\description{
Stepwise process of instantiating the Athena Vocabularies. Steps can be skipped using the \code{steps} argument though the order cannot be changed.
}
\seealso{
\code{\link[rlang]{parse_expr}}
\code{\link[pg13]{dc}},\code{\link[pg13]{is_conn_open}},\code{\link[pg13]{lsSchema}},\code{\link[pg13]{dropTable}},\code{\link[pg13]{send}}
\code{\link[cli]{cat_line}}
\code{\link[secretary]{typewrite}}
\code{\link[SqlRender]{render}}
}
