% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProcess.R
\name{runProcess}
\alias{runProcess}
\title{Run Update}
\usage{
runProcess(
  conn = conn,
  targetSchema = "public",
  cascade = TRUE,
  vocabularyPath,
  cpt4 = TRUE,
  verbose = TRUE
)
}
\description{
This function runs the entire process of dropping the target schema if it exists and populating the vocabulary tables.
}
