% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSetup.R
\name{runSetup}
\alias{runSetup}
\title{Run Full Setup}
\usage{
runSetup(
  conn = conn,
  targetSchema = "public",
  cascade = TRUE,
  vocabularyPath,
  cpt4 = TRUE,
  verbose = TRUE
)
}
\description{
This function runs the entire process of dropping the target schema if it exists and populating the vocabulary tables.
}
