% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vocabularyTableNames}
\alias{vocabularyTableNames}
\title{OMOP CDM 5.0 Vocabulary Tables}
\format{
A vector of length 11:
\describe{
\item{\code{vocabularyTableNames}}{character Table names}
}
}
\usage{
vocabularyTableNames
}
\description{
These are the names of the tables that comprise the OMOP Vocabulary Tables in the OMOP CDM v5.0.
}
\details{
DETAILS
}
\keyword{datasets}
