% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearSchema.R
\name{clearSchema}
\alias{clearSchema}
\title{Clear the Current Target Schema}
\usage{
clearSchema(conn, targetSchema, cascade = FALSE)
}
\arguments{
\item{conn}{A Connection object.}

\item{cascade}{If TRUE, will drop cascade the schema.}
}
\description{
Clear the Current Target Schema
}
