% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints}
\alias{constraints}
\title{Add Constraints}
\usage{
constraints(conn, target_schema, verbose = TRUE, render_sql = TRUE)
}
\description{
SQL can be found at \url{https://raw.githubusercontent.com/OHDSI/CommonDataModel/master/PostgreSQL/OMOP\%20CDM\%20postgresql\%20constraints.txt}
}
\seealso{
\code{\link[SqlRender]{render}}
\code{\link[pg13]{execute_n}}
}
