% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{ddl}
\alias{ddl}
\title{Execute Athena DDL}
\usage{
ddl(conn, targetSchema)
}
\description{
Execute Athena DDL
}
\seealso{
\code{\link[SqlRender]{render}},\code{\link[SqlRender]{readSql}}
\code{\link[pg13]{sourceFilePath}},\code{\link[pg13]{send}}
}
