% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{Execute Athena Indexes}
\usage{
indices(conn, targetSchema)
}
\description{
SQL for indices derived from \href{"https://raw.githubusercontent.com/OHDSI/CommonDataModel/master/PostgreSQL/OMOP\%20CDM\%20postgresql\%20pk\%20indexes.txt"}{OHDSI OMOP CDM PK Indexes.txt}
}
