% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRowCount.R
\name{logRowCount}
\alias{logRowCount}
\title{Log Table Row Counts}
\usage{
logRowCount(conn, targetSchema)
}
\description{
This function prints the number of rows for all the vocabulary tables in the R console. The results are also stored alongside a log of all previous data loads in a "setupAthena" cache subdirectory using the caching functions in the R.cache package.
}
