% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_args.R
\name{pkg_args}
\alias{pkg_args}
\title{setupAthena Arguments}
\arguments{
\item{path_to_csvs}{Path to folder containing the
unpacked vocabularies as csv files, downloaded as a zip
from \url{athena.ohdsi.org}.}

\item{umls_api_key}{UMLS API Key for CPT4 reconstitution.}

\item{conn}{A Postgres connection object from the
DatabaseConnector package. The user must either be a
superuser or a member of the pg_read_server_files role.}

\item{conn_fun}{Character string of the expression used
to connect to the target database for the setup. The user
must either be a superuser or a member of the
pg_read_server_files role.This option automatically
closes the connection on exit.}

\item{verbose}{If TRUE, prints back activity in the R
console as it is occurring.}

\item{render_sql}{If TRUE, the SQL statement is printed
back in the R console for reference.}

\item{target_schema}{Schema where the OMOP Vocabulary
Tables will be written to.}

\item{steps}{Options include c("prepare_cpt4",
"drop_tables","copy", "indices", "constraints", "log").
Each step corresponds to the function used and may be
independently executed.}

\item{release_version}{Release version as an exact string
provided vocabulary download link delivery via email.}
}
\description{
setupAthena Arguments
}
