% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpt4.R
\name{prepare_cpt4}
\alias{prepare_cpt4}
\title{Reconstitute CPT4 Concepts}
\usage{
prepare_cpt4(path_to_csvs, umls_api_key, verbose = TRUE)
}
\arguments{
\item{path_to_csvs}{Path to folder containing the unpacked vocabularies as csv files, downloaded as a zip from \url{athena.ohdsi.org}.}

\item{umls_api_key}{UMLS API Key for CPT4 reconstitution.}

\item{verbose}{If TRUE, prints back activity in the R console as it is occurring.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Run the java script that reconstitutes CPT4 concepts directly from the R console.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cli]{cat_line}}
\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{character(0)}}
\code{\link[purrr]{map}}
}
