% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_setup.R
\name{run_setup}
\alias{run_setup}
\title{Run Athena Setup}
\usage{
run_setup(
  conn,
  conn_fun,
  target_schema = "omop_vocabulary",
  steps = c("drop_tables", "copy", "indices", "constraints", "log"),
  path_to_csvs,
  umls_api_key,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{target_schema}{Schema where the OMOP Vocabulary ('Athena') Tables will be written to.}

\item{steps}{Options include c("prepare_cpt4", "drop_tables","copy", "indices", "constraints", "log"). Each step corresponds to the function used and may be independently executed. \code{\link{prepare_cpt4}} is not included in default because of its tendency to run slow and the java script used to reconstitute the CPT4 needs to run uninterrupted or else the resulting CONCEPT.csv file will be damaged. Therefore, it is advised to reconstitute CPT4 separately from the Command Line or using the aforementioned function in this package though the option to do it as part of \code{\link{run_setup}} is available.}

\item{path_to_csvs}{Path to folder containing the unpacked vocabularies as csv files, downloaded as a zip from \url{athena.ohdsi.org}.}

\item{umls_api_key}{UMLS API Key for CPT4 reconstitution.}
}
\value{
Updated OMOP Vocabulary (Athena) CONCEPT_ANCESTOR, CONCEPT_CLASS, CONCEPT_RELATIONSHIP, CONCEPT_SYNONYM, CONCEPT, DOMAIN, DRUG_STRENGTH, RELATIONSHIP, and VOCABULARY Tables in the given target schema.
}
\description{
Stepwise process of instantiating the Athena Vocabularies. Steps can be skipped using the \code{steps} argument though the order cannot be changed.
}
\seealso{
\code{\link[rlang]{parse_expr}}
\code{\link[pg13]{dc}},\code{\link[pg13]{is_conn_open}},\code{\link[pg13]{lsSchema}},\code{\link[pg13]{dropTable}},\code{\link[pg13]{send}}
\code{\link[cli]{cat_line}}
\code{\link[secretary]{typewrite}}
\code{\link[SqlRender]{render}}
}
