% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{runSetup}
\alias{runSetup}
\title{Run Full Setup}
\usage{
runSetup(
  conn = conn,
  targetSchema = "omop_vocabulary",
  cascade = TRUE,
  vocabularyPath,
  verbose = TRUE
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION, Default: conn}

\item{targetSchema}{PARAM_DESCRIPTION, Default: 'public'}

\item{cascade}{PARAM_DESCRIPTION, Default: TRUE}

\item{vocabularyPath}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function runs the entire process of dropping the target schema if it exists and populating the vocabulary tables.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[secretary]{typewrite_note}},\code{\link[secretary]{press_enter}},\code{\link[secretary]{typewrite}}
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{createSchema}}
\code{\link[crayon]{crayon}}
}
